refactor(persistence): separate observer persistence ports from SQLite
