+++ apps/observer/test/integration/session-commands.test.ts
@@ -22,7 +22,7 @@ import {
   createCommandQueue,
   createObserverCore,
   createObserverEventBus,
-  createObserverPersistence,
+  createSqliteObserverPersistence,
   openObserverSqlite,
   ProviderRegistry,
   registerObserverCommandHandlers,
@@ -1558,7 +1558,7 @@ function createFixture(
   const clock = { now: () => new Date(now) };
   const sqlite = openObserverSqlite({ clock });
   const ids = observerIds();
-  const persistence = createObserverPersistence({ sqlite, clock, idFactory: ids });
+  const persistence = createSqliteObserverPersistence({ sqlite, clock, idFactory: ids });
   const eventBus = createObserverEventBus();
   const queue = createCommandQueue({ persistence, clock, idFactory: ids, eventBus });
   const providers = new ProviderRegistry({
