+++ tests/diagnostics/debug-bundle/claude-provider.test.ts
@@ -7,7 +7,7 @@ import { writeDebugBundle } from "@station/observability";
 import {
   collectDiagnosticSnapshot,
   createObserverCore,
-  createObserverPersistence,
+  createSqliteObserverPersistence,
   openObserverSqlite,
   ProviderRegistry,
 } from "@station/observer/internal";
@@ -25,7 +25,7 @@ describe("Claude provider debug bundle diagnostics", () => {
 
     const clock = { now: () => new Date(now) };
     const sqlite = openObserverSqlite({ path: join(stateDir, "observer.sqlite"), clock });
-    const persistence = createObserverPersistence({
+    const persistence = createSqliteObserverPersistence({
       sqlite,
       clock,
       idFactory: ids(),
