+++ tests/diagnostics/debug-bundle/cleanup-command.test.ts
@@ -9,7 +9,7 @@ import {
   createObserverCore,
   createObserverEventBus,
   createObserverLogger,
-  createObserverPersistence,
+  createSqliteObserverPersistence,
   openObserverSqlite,
   ProviderRegistry,
   registerObserverCommandHandlers,
@@ -35,7 +35,7 @@ describe("cleanup command debug bundle diagnostics", () => {
     const clock = { now: () => new Date(now) };
     const sqlite = openObserverSqlite({ path: join(stateDir, "observer.sqlite"), clock });
     const ids = observerIds();
-    const persistence = createObserverPersistence({ sqlite, clock, idFactory: ids });
+    const persistence = createSqliteObserverPersistence({ sqlite, clock, idFactory: ids });
     const eventBus = createObserverEventBus();
     const logger = createObserverLogger({ stateDir, clock });
     const providers = new ProviderRegistry({
